res replaced use of printf() and relatives with csPrintf() and   relatives and use of sprintf() with use of csStrings. res fixed some 64bit warnings on VC7.0.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20794 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/hash.h
@@ -155,7 +155,7 @@ public:
    *
    * For a bigger list go to http://www.utm.edu/research/primes/
    */
-  csHash (int size = 23, int grow_rate = 5, int max_size = 20000)
+  csHash (size_t size = 23, size_t grow_rate = 5, size_t max_size = 20000)
     : Elements (size), Modulo (size), InitModulo (size),
       GrowRate (MIN (grow_rate, size)), MaxSize (max_size), Size (0)
   {
